requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==2.1.1
click==8.3.2
cryptography==46.0.7
Flask==2.2.5
google-auth==2.49.2
google-genai==0.1.0
humandelta==0.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==11.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==0.21.0
requests==2.28.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
websockets==14.2
Werkzeug==2.2.3
